prepared map for avx

+++ src/ion/modules/map/functions/map_init.h
@@ -3,8 +3,8 @@ static inline
 #endif
 
 struct map map_init (
-    u64 key_typesize,
-    u64 value_typesize,
-    u64 initial_capacity,
+    u32 key_typesize,
+    u32 value_typesize,
+    u32 initial_capacity,
     struct memory* allocator
 );
